+++ enterprise/server/CMakeLists.txt
@@ -1,21 +1,22 @@
 sourcemeta_library(NAMESPACE sourcemeta PROJECT one NAME enterprise_server
-  PRIVATE_HEADERS actions.h
-  SOURCES action_mcp_v1.cc action_jsonschema_rdf_v1.cc)
+  PRIVATE_HEADERS action_mcp_v1.h action_jsonschema_rdf_v1.h)
 
-target_compile_definitions(sourcemeta_one_enterprise_server PUBLIC SOURCEMETA_ONE_ENTERPRISE)
+target_compile_definitions(sourcemeta_one_enterprise_server INTERFACE SOURCEMETA_ONE_ENTERPRISE)
 
-target_link_libraries(sourcemeta_one_enterprise_server PUBLIC sourcemeta::one::router)
-target_link_libraries(sourcemeta_one_enterprise_server PUBLIC sourcemeta::one::http)
-target_link_libraries(sourcemeta_one_enterprise_server PUBLIC sourcemeta::one::search)
-target_link_libraries(sourcemeta_one_enterprise_server PUBLIC sourcemeta::core::uritemplate)
-target_link_libraries(sourcemeta_one_enterprise_server PUBLIC sourcemeta::blaze::evaluator)
-target_link_libraries(sourcemeta_one_enterprise_server PRIVATE sourcemeta::core::json)
-target_link_libraries(sourcemeta_one_enterprise_server PRIVATE sourcemeta::core::jsonpointer)
-target_link_libraries(sourcemeta_one_enterprise_server PRIVATE sourcemeta::core::jsonrpc)
-target_link_libraries(sourcemeta_one_enterprise_server PRIVATE sourcemeta::core::uri)
-target_link_libraries(sourcemeta_one_enterprise_server PRIVATE sourcemeta::one::shared)
-target_link_libraries(sourcemeta_one_enterprise_server PRIVATE sourcemeta::core::mcp)
-target_link_libraries(sourcemeta_one_enterprise_server PRIVATE sourcemeta::core::text)
-target_link_libraries(sourcemeta_one_enterprise_server PRIVATE sourcemeta::one::metapack)
-target_link_libraries(sourcemeta_one_enterprise_server PRIVATE sourcemeta::blaze::output)
-target_link_libraries(sourcemeta_one_enterprise_server PRIVATE sourcemeta::core::jsonld)
+target_link_libraries(sourcemeta_one_enterprise_server INTERFACE sourcemeta::one::router)
+target_link_libraries(sourcemeta_one_enterprise_server INTERFACE sourcemeta::one::http)
+target_link_libraries(sourcemeta_one_enterprise_server INTERFACE sourcemeta::one::search)
+target_link_libraries(sourcemeta_one_enterprise_server INTERFACE sourcemeta::core::uritemplate)
+target_link_libraries(sourcemeta_one_enterprise_server INTERFACE sourcemeta::blaze::evaluator)
+target_link_libraries(sourcemeta_one_enterprise_server INTERFACE sourcemeta::core::json)
+target_link_libraries(sourcemeta_one_enterprise_server INTERFACE sourcemeta::core::jsonpointer)
+target_link_libraries(sourcemeta_one_enterprise_server INTERFACE sourcemeta::core::jsonrpc)
+target_link_libraries(sourcemeta_one_enterprise_server INTERFACE sourcemeta::core::uri)
+target_link_libraries(sourcemeta_one_enterprise_server INTERFACE sourcemeta::one::shared)
+target_link_libraries(sourcemeta_one_enterprise_server INTERFACE sourcemeta::core::mcp)
+target_link_libraries(sourcemeta_one_enterprise_server INTERFACE sourcemeta::core::text)
+target_link_libraries(sourcemeta_one_enterprise_server INTERFACE sourcemeta::one::metapack)
+target_link_libraries(sourcemeta_one_enterprise_server INTERFACE sourcemeta::blaze::output)
+target_link_libraries(sourcemeta_one_enterprise_server INTERFACE sourcemeta::core::jsonld)
+target_link_libraries(sourcemeta_one_enterprise_server INTERFACE sourcemeta::core::numeric)
+target_link_libraries(sourcemeta_one_enterprise_server INTERFACE sourcemeta::core::http)
