+++ enterprise/server/include/sourcemeta/one/enterprise_server.h
@@ -1,6 +1,7 @@
 #ifndef SOURCEMETA_ONE_ENTERPRISE_SERVER_H_
 #define SOURCEMETA_ONE_ENTERPRISE_SERVER_H_
 
-#include <sourcemeta/one/enterprise_server_actions.h>
+#include <sourcemeta/one/enterprise_server_action_jsonschema_rdf_v1.h>
+#include <sourcemeta/one/enterprise_server_action_mcp_v1.h>
 
 #endif
